rednote_research/pyproject.toml
[project]
name = "rednote-research"
version = "0.1.0"
description = "基于MCP的小红书深度研究智能体"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.0.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "sse-starlette>=1.0.0",
    "uvicorn>=0.20.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
rednote-research = "rednote_research.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
